(#74) Adjust Hyper-V provider config
The hyper-v box can potentially have the max memory set lower than the
desired amount of memory in the test environment Vagrantfile. Overring
the max memory with `nil` enables any memory amount to be used.

In testing, it also took longer than expected to for the box to get an
IP address, and so the IP Address Timeout has been increased to give
more time for this step to complete.