(#74) Update README with Hyper-V info
Add a section on using Hyper-V instead of VirtualBox, including a
recommendation to prefer VirtualBox and linking to known limitations.

This also updates the table of contents to include the newly added
section as well as two sections added in a previous commit that were
not added to the list.